graphql-user/requirements.txt
fastapi==0.78.0
pydantic==1.9.1
strawberry-graphql==0.114.2
